Rename `hankel_matrix.py` to `hankel_matrix_utils.py`
- Rename the module to prevent name conflicts with the
  `hankel_matrix()` function within it.
- Update docs to align with the module rename.

This ensures correct docs building with Sphinx and avoids confusion
for users.
